+++ docs/metro/js/metro.js
@@ -100,8 +100,8 @@ var isTouch = (('ontouchstart' in window) || (navigator.MaxTouchPoints > 0) || (
 
 var Metro = {
 
-    version: "4.2.34-dev [16:49 4-0-2019]",
-    versionFull: "4.2.34-dev [16:49 4-0-2019]",
+    version: "4.2.34-dev [17:39 4-0-2019]",
+    versionFull: "4.2.34-dev [17:39 4-0-2019]",
     isTouchable: isTouch,
     fullScreenEnabled: document.fullscreenEnabled,
     sheet: null,
@@ -3419,8 +3419,10 @@ var Utils = {
     isEmbedObject: function(val){
         var embed = ["iframe", "object", "embed", "video"];
         var result = false;
-        $.each(embed, function(){
-            if (val.indexOf(this) !== -1) {
+        $.each(embed, function(i, v){
+            if (typeof val === "string" && val.toLowerCase() === v) {
+                result = true;
+            } else if (val.nodeType !== undefined && val.tagName.toLowerCase() === v) {
                 result = true;
             }
         });
@@ -11780,6 +11782,7 @@ var Input = {
     },
     options: {
         autocomplete: null,
+        autocompleteDivider: ",",
         autocompleteListHeight: 200,
 
         history: false,
@@ -11943,7 +11946,7 @@ var Input = {
             if (autocomplete_obj !== false) {
                 that.autocomplete = autocomplete_obj;
             } else {
-                this.autocomplete = Utils.strToArray(o.autocomplete);
+                this.autocomplete = Utils.strToArray(o.autocomplete, o.autocompleteDivider);
             }
             $("<div>").addClass("autocomplete-list").css({
                 maxHeight: o.autocompleteListHeight,
@@ -12067,15 +12070,15 @@ var Input = {
                 display: items.length > 0 ? "block" : "none"
             });
 
-            $.each(items, function(){
-                var index = this.toLowerCase().indexOf(val);
-                var item = $("<div>").addClass("item").attr("data-autocomplete-value", this);
+            $.each(items, function(i, v){
+                var index = v.toLowerCase().indexOf(val);
+                var item = $("<div>").addClass("item").attr("data-autocomplete-value", v);
                 var html;
 
                 if (index === 0) {
-                    html = "<strong>"+this.substr(0, val.length)+"</strong>"+this.substr(val.length);
+                    html = "<strong>"+v.substr(0, val.length)+"</strong>"+v.substr(val.length);
                 } else {
-                    html = this.substr(0, index) + "<strong>"+this.substr(index, val.length)+"</strong>"+this.substr(index + val.length);
+                    html = v.substr(0, index) + "<strong>"+v.substr(index, val.length)+"</strong>"+v.substr(index + val.length);
                 }
                 item.html(html).appendTo(autocompleteList);
             })